# openapi-spec.yaml
openapi: "3.0.3"
info:
  version: 0.0.1
  title: Doggoschool API
  description: A small sample API to grasp the basics of the OpenAPI Specification
servers:
  - url: http://localhost:8080/api
paths:
  /dogs:
    get:
      summary: List all doggos
      operationId: getDogs
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: An paged array of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DogDTO"
    post:
      summary: introduce a new doggo to the doggos
      operationId: addDog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DogDTO"
      responses:
        201:
          description: added the doggo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DogDTO"
  /dogs/{dogId}:
    patch:
      summary: change owner of doggo
      operationId: changeDogOwner
      parameters:
        - name: dogId
          in: path
          description: ID to change owner of dog
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerDTO"

      responses:
        204:
          description: owner succesfully changed
components:
  schemas:
    DogDTO:
      required:
        - name
        - breed
        - food
        - age
      properties:
        id:
          type: integer
          example: 42
          minimum: 0
          readOnly: true
        name:
          type: string
          minLength: 1
          example: Kira
        breed:
          type: string
          minLength: 1
          example: Golden Retriever
        owner:
          $ref: "#/components/schemas/OwnerDTO"
        food:
          type: string
          minLength: 1
          example: t-bone Steak
        age:
          type: integer
          minLength: 1
          example: 7
    OwnerDTO:
      required:
        - name
        - age
      properties:
        id:
          type: integer
          example: 42
          minimum: 0
          readOnly: true
        name:
          type: string
          minLength: 1
          example: Kira
        age:
          type: number
          minLength: 1
          example: 18
    FoodDTO:
      required:
        - type
        - mainIngredients
        - protein
      properties:
        id:
          type: integer
          example: 42
          minimum: 0
          readOnly: true
        type:
          type: string
          minLength: 1
          example: meat
        mainIngredients:
          type: string
          minLength: 1
          example: chicken
        protein:
          type: string
          minLength: 1
          example: eiweiss